Exercise 5 - v1.1
The exercise to modify files in repository then observ it from local machine. And it would be nice to have example with git ammend, merge and/or rebase in the fork and merge to upstream eventually.
So far, the file is created in forked repo which also behind to the upstream. After commit this change and push to the forked repo, we can rebase on newly fetched upstream. A line feed on the -m message would convinient for git log --oneline. Now we are trying to send git hub pull request to the upstream owner. This requires to download the github hub program